requirements.txt
python-telegram-bot[job-queue]==20.8
paramiko
cryptography
jdatetime
matplotlib
requests
psycopg2-binary
websockets
psutil
